Improve function name in example
It was referring to a module that creates the role and returns a function name, but it is a bad example because you don't need to use that module when using this module. You can if you want, but it's not necessary, so it makes a bad example.

The module was https://gitlab.com/claranet-pcp/terraform/aws/terraform-aws-lambda-role which ensures that policies are attached before the function name gets returned, helping to ensure that the role is ready before the function is created.